feature： 封装组件
