+++ styles/globals.css
@@ -150,13 +150,13 @@ nav {
 }
 
 .glassmorphism{
-  background: rgba(255, 255, 255, 0.2);
-  -webkit-backdrop-filter: blur(3px);
-  backdrop-filter: blur(3px);
+  background: hsla(0, 0%, 100%, .75);
+  -webkit-backdrop-filter: blur(5px);    
+  backdrop-filter: blur(5px);   
 }
 
 .dark .glassmorphism{
-  background: rgba(31, 41, 55, 0.2);
+  background: rgba(31, 41, 55, .75);
   -webkit-backdrop-filter: blur(3px);
   backdrop-filter: blur(3px);
 }
